+++ tests/InflectorTest.php
@@ -160,6 +160,15 @@ final class InflectorTest extends TestCase
         $this->assertEquals('ひらがなHepimiz', $inflector->toCamelCase('ひらがな_hepimiz'));
     }
 
+    public function testCamelCaseToSnakeCase(): void
+    {
+        $inflector = new Inflector();
+
+        $this->assertEquals('user_name', $inflector->camelCaseToSnakeCase('userName'));
+        $this->assertEquals('ひらがな_hepimiz', $inflector->camelCaseToSnakeCase('ひらがなHepimiz'));
+        $this->assertEquals('travel_s_guide', $inflector->camelCaseToSnakeCase('travelSGuide'));
+    }
+
     public function testToTable(): void
     {
         $inflector = new Inflector();